refactor(web): extract newsletter cli adapters from app
