+++ web/newsletter_clients.py
@@ -1,0 +1,449 @@
+# mypy: disable-error-code="no-untyped-def,assignment"
+# flake8: noqa
+"""Newsletter CLI adapters used by the Flask web app."""
+
+import datetime
+import logging
+import os
+import sys
+
+try:
+    from runtime_paths import resolve_env_file_path, resolve_project_root
+except ImportError:
+    from web.runtime_paths import (  # pragma: no cover
+        resolve_env_file_path,
+        resolve_project_root,
+    )
+
+from newsletter_core.public.generation import (
+    GenerateNewsletterRequest,
+    NewsletterGenerationError,
+)
+from newsletter_core.public.generation import (
+    generate_newsletter as generate_newsletter_core,
+)
+
+
+class RealNewsletterCLI:
+    def __init__(self):
+        # CLI 경로 설정 - dev에서는 프로젝트 루트, PyInstaller에서는 bundle 루트
+        self.project_root = resolve_project_root()
+        # 실행 결과물(output, storage)은 실행 파일 디렉토리에 남기도록 고정
+        self.runtime_work_dir = os.path.dirname(resolve_env_file_path())
+        self.module_root = self.project_root
+        self.timeout = 900  # 15분 타임아웃으로 증가
+
+        # 환경 확인
+        self._check_environment()
+
+    def _check_environment(self):
+        """환경 설정 확인"""
+        # 프로젝트 루트 확인
+        module_candidates = [os.path.join(self.project_root, "newsletter")]
+        if getattr(sys, "frozen", False) and hasattr(sys, "_MEIPASS"):
+            module_candidates.insert(0, os.path.join(sys._MEIPASS, "newsletter"))
+
+        module_path = next((p for p in module_candidates if os.path.exists(p)), None)
+        if not module_path:
+            raise Exception(f"Newsletter module not found in {self.project_root}")
+
+        self.module_root = os.path.dirname(module_path)
+
+        # .env 파일 확인
+        env_candidates = [
+            resolve_env_file_path(),
+            os.path.join(self.project_root, ".env"),
+        ]
+        env_file = next(
+            (p for p in env_candidates if os.path.exists(p)), env_candidates[0]
+        )
+        if not os.path.exists(env_file):
+            print(f"⚠️  Warning: .env file not found at {env_file}")
+            print("⚠️  This may cause longer processing times or fallback to mock mode")
+
+        # API 키 확인
+        api_keys_status = self._check_api_keys()
+
+        print("✅ Environment check passed")
+        print(f"   Project root: {self.project_root}")
+        print(f"   Runtime work dir: {self.runtime_work_dir}")
+        print(f"   Newsletter module exists: {os.path.exists(module_path)}")
+        print(f"   .env file exists: {os.path.exists(env_file)}")
+        print(f"   API keys configured: {api_keys_status}")
+
+    def _check_api_keys(self):
+        """API 키 설정 상태 확인"""
+        required_keys = {
+            "GEMINI_API_KEY": "Gemini (primary LLM)",
+            "OPENAI_API_KEY": "OpenAI (fallback LLM)",
+            "POSTMARK_SERVER_TOKEN": "Email service",  # nosec B105
+        }
+
+        configured = []
+        missing = []
+
+        for key, description in required_keys.items():
+            if os.getenv(key):
+                configured.append(f"{description} ✅")
+            else:
+                missing.append(f"{description} ❌")
+
+        if missing:
+            print(f"⚠️  Missing API keys: {', '.join(missing)}")
+            print("   This may cause slower performance or feature limitations")
+
+        return f"{len(configured)}/{len(required_keys)} configured"
+
+    def generate_newsletter(
+        self,
+        keywords=None,
+        domain=None,
+        template_style="compact",
+        email_compatible=False,
+        period=14,
+    ):
+        """공유 코어 API를 사용하여 뉴스레터 생성"""
+        try:
+            request_payload = GenerateNewsletterRequest(
+                keywords=keywords,
+                domain=domain,
+                template_style=template_style,
+                email_compatible=email_compatible,
+                period=period,
+            )
+            result = generate_newsletter_core(request_payload)
+
+            return {
+                "content": result["html_content"],
+                "title": result["title"],
+                "status": result["status"],
+                "generation_stats": result.get("generation_stats", {}),
+                "input_params": result.get("input_params", {}),
+                "error": result.get("error"),
+            }
+
+        except NewsletterGenerationError as exc:
+            logging.error("Core generation failed: %s", exc)
+            return self._fallback_response(keywords or domain, str(exc))
+        except Exception as e:
+            error_msg = f"Unexpected error: {str(e)}"
+            logging.error(error_msg, exc_info=True)
+            return self._fallback_response(keywords or domain, error_msg)
+
+    def _find_latest_html_file(self, output_dir, keywords=None):
+        """output 디렉토리에서 최신 HTML 파일 찾기"""
+        try:
+            if not os.path.exists(output_dir):
+                logging.error(f"Output directory does not exist: {output_dir}")
+                return None
+
+            html_files = [f for f in os.listdir(output_dir) if f.endswith(".html")]
+            if not html_files:
+                logging.error(f"No HTML files found in {output_dir}")
+                return None
+
+            logging.info(f"Found {len(html_files)} HTML files in {output_dir}")
+
+            # 키워드가 있으면 해당 키워드가 포함된 파일을 우선적으로 찾기
+            if keywords:
+                keyword_str = (
+                    keywords if isinstance(keywords, str) else ",".join(keywords)
+                )
+                keyword_files = [
+                    f
+                    for f in html_files
+                    if any(
+                        kw.strip().lower() in f.lower() for kw in keyword_str.split(",")
+                    )
+                ]
+                if keyword_files:
+                    html_files = keyword_files
+                    logging.info(
+                        f"Filtered to {len(keyword_files)} files matching keywords: {keyword_str}"
+                    )
+
+            # 최신 파일 찾기 (생성 시간 기준)
+            latest_file = max(
+                html_files, key=lambda x: os.path.getctime(os.path.join(output_dir, x))
+            )
+            file_path = os.path.join(output_dir, latest_file)
+
+            logging.info(f"Reading latest HTML file: {latest_file}")
+            logging.info(f"File path: {file_path}")
+            logging.info(f"File size: {os.path.getsize(file_path)} bytes")
+
+            # 여러 인코딩으로 시도
+            encodings = ["utf-8", "utf-8-sig", "cp949", "euc-kr", "latin1"]
+
+            for encoding in encodings:
+                try:
+                    with open(file_path, "r", encoding=encoding) as f:
+                        content = f.read()
+                        logging.info(f"Successfully read file with {encoding} encoding")
+                        logging.info(f"Content length: {len(content)} characters")
+                        return content
+                except UnicodeDecodeError:
+                    logging.warning(f"Failed to read with {encoding} encoding")
+                    continue
+
+            logging.error("Failed to read file with any encoding")
+            return None
+
+        except Exception as e:
+            logging.error(f"Error reading HTML file: {e}")
+            return None
+
+    def _extract_title_from_html(self, html_content):
+        """HTML에서 제목 추출"""
+        try:
+            import re
+
+            title_match = re.search(
+                r"<title>(.*?)</title>", html_content, re.IGNORECASE
+            )
+            if title_match:
+                return title_match.group(1)
+            return None
+        except Exception:
+            return None
+
+    def _extract_generation_stats(self, cli_output):
+        """CLI 출력에서 생성 통계 정보 추출"""
+        stats = {}
+        try:
+            import re
+
+            # 처리 시간 추출
+            time_pattern = r"(\w+)\s+time:\s*([\d.]+)\s*seconds"
+            time_matches = re.findall(time_pattern, cli_output)
+            if time_matches:
+                stats["step_times"] = {step: float(time) for step, time in time_matches}
+
+            # 총 시간 추출
+            total_time_pattern = r"Total generation time:\s*([\d.]+)\s*seconds"
+            total_match = re.search(total_time_pattern, cli_output)
+            if total_match:
+                stats["total_time"] = float(total_match.group(1))
+
+            # 기사 수 추출
+            article_pattern = r"(\d+)\s+articles?\s+(?:collected|processed|found)"
+            article_match = re.search(article_pattern, cli_output, re.IGNORECASE)
+            if article_match:
+                stats["articles_count"] = int(article_match.group(1))
+
+            # 키워드 정보 추출
+            if "Generated keywords:" in cli_output:
+                keyword_pattern = r"Generated keywords:\s*(.+)"
+                keyword_match = re.search(keyword_pattern, cli_output)
+                if keyword_match:
+                    stats["generated_keywords"] = keyword_match.group(1).strip()
+
+        except Exception as e:
+            logging.warning(f"Failed to extract generation stats: {e}")
+
+        return stats
+
+    def _fallback_response(self, input_param, error_msg):
+        """에러 발생 시 폴백 응답 생성"""
+        title = f"Newsletter Generation Failed: {input_param}"
+        content = f"""
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>{title}</title>
+    <style>
+        body {{ font-family: Arial, sans-serif; margin: 20px; line-height: 1.6; }}
+        .error {{ background: #f8d7da; color: #721c24; padding: 20px; border-radius: 5px; border: 1px solid #f5c6cb; }}
+        .info {{ background: #d4edda; color: #155724; padding: 15px; border-radius: 5px; border: 1px solid #c3e6cb; margin-top: 20px; }}
+    </style>
+</head>
+<body>
+    <div class="error">
+        <h2>Newsletter Generation Error</h2>
+        <p><strong>Input:</strong> {input_param}</p>
+        <p><strong>Error:</strong> {error_msg}</p>
+        <p><strong>Time:</strong> {datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S')}</p>
+    </div>
+    <div class="info">
+        <h3>What happened?</h3>
+        <p>The newsletter CLI encountered an error during generation. This could be due to:</p>
+        <ul>
+            <li>Network connectivity issues</li>
+            <li>API rate limits or quota exceeded</li>
+            <li>Invalid input parameters</li>
+            <li>CLI timeout (process took longer than {self.timeout} seconds)</li>
+        </ul>
+        <p>Please try again with different parameters or check the system logs for more details.</p>
+    </div>
+</body>
+</html>
+"""
+
+        return {
+            "content": content,
+            "title": title,
+            "status": "error",
+            "error": error_msg,
+        }
+
+
+# MockNewsletterCLI는 폴백용으로 유지
+class MockNewsletterCLI:
+    def __init__(self):
+        pass
+
+    def generate_newsletter(
+        self,
+        keywords=None,
+        domain=None,
+        template_style="compact",
+        email_compatible=False,
+        period=14,
+    ):
+        """Mock newsletter generation with more realistic content"""
+        if keywords:
+            if isinstance(keywords, list):
+                keyword_list = [k.strip() for k in keywords]
+            else:
+                keyword_list = [k.strip() for k in keywords.split(",")]
+            title = f"Newsletter: {', '.join(keyword_list)} (Mock)"
+
+            content = f"""
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>{title}</title>
+    <style>
+        body {{ font-family: Arial, sans-serif; margin: 20px; line-height: 1.6; }}
+        .header {{ background: #f4f4f4; padding: 20px; text-align: center; }}
+        .content {{ padding: 20px; }}
+        .article {{ margin-bottom: 20px; padding: 15px; border-left: 3px solid #007bff; }}
+        .article h3 {{ color: #007bff; margin-top: 0; }}
+        .footer {{ background: #f4f4f4; padding: 15px; text-align: center; font-size: 0.9em; }}
+        .mock-notice {{ background: #fff3cd; color: #856404; padding: 15px; border-radius: 5px; border: 1px solid #ffeaa7; margin-bottom: 20px; }}
+    </style>
+</head>
+<body>
+    <div class="mock-notice">
+        <strong>⚠️ Mock Mode:</strong> This is a test newsletter generated using mock data.
+        Template Style: {template_style} | Email Compatible: {email_compatible} | Period: {period} days
+    </div>
+    <div class="header">
+        <h1>{title}</h1>
+        <p>Generated on {datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S')}</p>
+    </div>
+    <div class="content">
+        <h2>Top Stories</h2>
+"""
+
+            for i, keyword in enumerate(keyword_list, 1):
+                content += f"""
+        <div class="article">
+            <h3>{i}. Latest developments in {keyword}</h3>
+            <p>This is a mock article about recent developments in {keyword}. In a real implementation, this would contain actual news content collected from various sources.</p>
+            <p><strong>Key points:</strong></p>
+            <ul>
+                <li>Market trends and analysis for {keyword}</li>
+                <li>Recent technological advancements</li>
+                <li>Industry impact and future outlook</li>
+            </ul>
+        </div>
+"""
+
+            content += """
+    </div>
+    <div class="footer">
+        <p>This newsletter was generated by Newsletter Generator Web Service (Mock Mode)</p>
+        <p>For more information, visit our website</p>
+    </div>
+</body>
+</html>
+"""
+
+        elif domain:
+            title = f"Newsletter: {domain} Domain Insights (Mock)"
+            content = f"""
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>{title}</title>
+    <style>
+        body {{ font-family: Arial, sans-serif; margin: 20px; line-height: 1.6; }}
+        .header {{ background: #f4f4f4; padding: 20px; text-align: center; }}
+        .content {{ padding: 20px; }}
+        .section {{ margin-bottom: 30px; }}
+        .section h3 {{ color: #007bff; border-bottom: 2px solid #007bff; padding-bottom: 5px; }}
+        .footer {{ background: #f4f4f4; padding: 15px; text-align: center; font-size: 0.9em; }}
+        .mock-notice {{ background: #fff3cd; color: #856404; padding: 15px; border-radius: 5px; border: 1px solid #ffeaa7; margin-bottom: 20px; }}
+    </style>
+</head>
+<body>
+    <div class="mock-notice">
+        <strong>⚠️ Mock Mode:</strong> This is a test newsletter generated using mock data.
+        Template Style: {template_style} | Email Compatible: {email_compatible} | Period: {period} days
+    </div>
+    <div class="header">
+        <h1>{title}</h1>
+        <p>Generated on {datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S')}</p>
+    </div>
+    <div class="content">
+        <div class="section">
+            <h3>Recent Developments in {domain}</h3>
+            <p>Mock analysis of recent developments in the {domain} sector. This would contain real news and insights in the actual implementation.</p>
+        </div>
+        <div class="section">
+            <h3>Key Trends</h3>
+            <p>Analysis of key trends affecting the {domain} industry, including market movements and technological advances.</p>
+        </div>
+        <div class="section">
+            <h3>Future Outlook</h3>
+            <p>Forward-looking analysis and predictions for the {domain} sector based on current trends and market indicators.</p>
+        </div>
+    </div>
+    <div class="footer">
+        <p>This newsletter was generated by Newsletter Generator Web Service (Mock Mode)</p>
+    </div>
+</body>
+</html>
+"""
+        else:
+            title = "Sample Newsletter (Mock)"
+            content = """
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Sample Newsletter (Mock)</title>
+</head>
+<body>
+    <h1>Sample Newsletter (Mock)</h1>
+    <p>No keywords or domain specified. Please provide either keywords or a domain for newsletter generation.</p>
+</body>
+</html>
+"""
+
+        return {"content": content, "title": title, "status": "success"}
+
+
+def create_newsletter_cli():
+    """Create real CLI adapter, falling back to mock adapter on failure."""
+    try:
+        cli = RealNewsletterCLI()
+        print("✅ Using RealNewsletterCLI for actual newsletter generation")
+        print(f"   Project root: {cli.project_root}")
+        print(f"   Timeout: {cli.timeout} seconds")
+        return cli
+    except Exception as e:
+        print(f"❌ Failed to initialize RealNewsletterCLI: {e}")
+        import traceback
+
+        traceback.print_exc()
+        print("⚠️  Falling back to MockNewsletterCLI")
+        return MockNewsletterCLI()
